% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impartial.R
\name{impartialReg}
\alias{impartialReg}
\alias{lm_impartial}
\alias{predict.impartialLM}
\title{Creating Impartial Regression Estimates}
\usage{
lm_impartial(data, colList)

\method{predict}{impartialLM}(object, newdata = NULL)
}
\arguments{
\item{data}{matrix or data.frame of covariates.}

\item{colList}{A list of named column specifications. List elements can
either be column numbers or column names. The data frame merely needs to be
able to be subset according to an element of this list. colList includes
some subset of c("Y","S","X","W"). An element named Y is always required,
and must have S if include W.}

\item{object}{an object of class impartialLM; expected to be the list output
from the lm_impartial function.}

\item{newdata}{an optional matrix or data frame in which to look for
variables with which to predict. If omitted, the fitted values are used.}
}
\value{
A list which includes the following components:
  \item{coeffWS}{coefficients from W~S+X; included in final output only if W
  is included.} \item{coeffX}{coefficients from Y~S+X+W.} \item{imEst}{impartial
  estimates given colList.} \item{colList}{covariate groupings.}
  \item{dataNames}{column names after transforming data; used for checking in
  predict.impartialLM.}
}
\description{
The function lm_impartial takes a data frame or matrix as well
  as a list specifying which columns correspond to various covariate groups:
  response, sensitive, legitimate, and suspect. It then produces estimates of
  the response which are impartial with respect to the sensitive covariates.
  Note that W can include estimates from other models which are not
  constrained to be fair or impartial.
}
\details{
The order of the variables in the model is very important:
  lm(Y~S+X+W) vs lm(Y~S+W+X). R will keep the first variables and drop the
  latter if they are collinear. Therefore, if W = S + X, X is kept in the
  first model while W is kept in the second. This can have a large impact on
  the coefficient for S. In fairness cases, it is better to keep X, since
  this is treated as legitimate variability, while W wont be. Therefore use
  Y~S+X+W. For the predict function, the data from which to predict does not
  include an intercept column. Before computing predictions on newdata, a
  check for equivalent model matrices is made. This requires column names in
  newdata to match those of the original data. Factors columns also need to
  include levels for all categories else equivalent model matrices will not
  be created.
}
\examples{
  data(swiss)
  colList1 = list("Y" = 1, "S" = 5, "X" = setdiff(1:6, c(1,5)))
  lmOut1 = lm_impartial(swiss, colList1)
  colList2 = list("Y" = 1, "S" = 5, "X" = setdiff(1:6, c(1,4,5)), "W" = 4)
  lmOut2 = lm_impartial(swiss, colList2)
}
